requirements.txt
numpy
pandas
pyarrow
pillow
torch
torchvision
timm
peft
